+++ pkgs/test/test/runner/browser/runner_test.dart
@@ -700,10 +700,9 @@ void main() {
     await d.file('test.dart', _failure).create();
 
     var test = await runTest(['-p', 'chrome', '--verbose-trace', 'test.dart']);
-    expect(
-        test.stdout,
-        containsInOrder(
-            [' main.<fn>', 'package:test', 'dart:async/zone.dart']));
+    expect(test.stdout,
+        containsInOrder([' main.<fn>', 'package:test', 'dart:async/zone.dart']),
+        skip: 'https://github.com/dart-lang/sdk/issues/41949');
     await test.shouldExit(1);
   }, tags: 'chrome');