+++ pkgs/test_api/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 0.2.16-dev
 
+* Deprecate `LiveTestController.liveTest`, the `LiveTestController` instance now
+  implements `LiveTest` and can be used directly.
+
 ## 0.2.15
 
 * Cancel any StreamQueue that is created as a part of a stream matcher once it