+++ pkgs/test_api/CHANGELOG.md
@@ -7,6 +7,9 @@
     group now has the empty string as its name.
 * Add `languageVersionComment` on the `MetaData` class. This should only be
   presen for test suites.
+* Deprecate `LiveTestController.liveTest`, the `LiveTestController` instance now
+  implements `LiveTest` and can be used directly.
+
 
 ## 0.2.15